deploy/scripts/persistent/UintListRegistry.ts
import type { UintListRegistryArgs } from '@enzymefinance/protocol';
import type { DeployFunction } from 'hardhat-deploy/types';

const fn: DeployFunction = async function (hre) {
  const {
    deployments: { deploy, get },
    ethers: { getSigners },
  } = hre;

  const deployer = (await getSigners())[0];
  const dispatcher = await get('Dispatcher');

  await deploy('UintListRegistry', {
    args: [dispatcher.address] as UintListRegistryArgs,
    from: deployer.address,
    log: true,
    skipIfAlreadyDeployed: true,
  });
};

fn.tags = ['Persistent', 'UintListRegistry'];
fn.dependencies = ['Dispatcher'];

export default fn;
